MAINT: Makes compatible with latest pandas (#112)
* MAINT: Makes compatible with latest pandas
Fixes #111 Pandas changed the reduce functionality
http://pandas.pydata.org/pandas-docs/stable/whatsnew.html?highlight=rolling_mean#changes-to-make-output-of-dataframe-apply-consistent
the new api is backwards incompatible with >0.23.4. The changes in this
pull request work with both versions of pandas. There may be a better
way to achive this, though I think it is preferable to keep
sourcetracker compatible with the recent versions of pandas

* Added the replace flag back in

* Added the replace flag back in

* updated changelog