Issue 106 (#120)
* ENH: Replaces ipyparallel with mutlithreading
fixes #106, fixes #98, fixes #92
This pull request replaces ipyparallel with the python multithreading
library. This should prevent some of the issues that we have seen
running sourcetracker in parallel previously

* fixed typo

* Reduces duplicated code

* Updated doctstring example to pass

* removed reference to ipyparallel

* ENH: Replaces ipyparallel with multithreading (#117)

* ENH: Replaces ipyparallel with mutlithreading
fixes #106, fixes #98, fixes #92
This pull request replaces ipyparallel with the python multithreading
library. This should prevent some of the issues that we have seen
running sourcetracker in parallel previously

* fixed typo

* Reduces duplicated code

* Updated doctstring example to pass

* removed reference to ipyparallel